cameraserver: Fix floating point truncation for getUHRMaxJpegBufferSize
Bug: 238103879

Test: Vendor testing
Test: DngCreatorTest.java on cuttlefish

Merged-In: I6553ed5f4b617e70516cfaac3c248da3176912da

+++ services/camera/libcameraservice/utils/SessionConfigurationUtils.cpp
@@ -74,7 +74,7 @@ camera3::Size SessionConfigurationUtils::getMaxJpegResolution(const CameraMetada
 
 size_t SessionConfigurationUtils::getUHRMaxJpegBufferSize(camera3::Size uhrMaxJpegSize,
         camera3::Size defaultMaxJpegSize, size_t defaultMaxJpegBufferSize) {
-    return (uhrMaxJpegSize.width * uhrMaxJpegSize.height) /
+    return ((float)uhrMaxJpegSize.width * uhrMaxJpegSize.height) /
             (defaultMaxJpegSize.width * defaultMaxJpegSize.height) * defaultMaxJpegBufferSize;
 }
